refactor(mobile-editor): extract header/stage/controls
Slim the mobile-editor index into a thin orchestrator; split the header
(close/index/saving/menu), stage (face editor) and controls (prev/flip/
next) into colocated components that each inject the editor composable.
